+++ modules/dxp/apps/oauth/oauth-service/src/main/java/com/liferay/oauth/internal/V10aOAuth.java
@@ -15,6 +15,7 @@
 package com.liferay.oauth.internal;
 
 import com.liferay.oauth.configuration.OAuthConfigurationValues;
+import com.liferay.oauth.internal.listener.V10aOAuthDebugCacheListener;
 import com.liferay.oauth.model.OAuthApplication;
 import com.liferay.oauth.model.OAuthUser;
 import com.liferay.oauth.service.OAuthApplicationLocalServiceUtil;
@@ -28,7 +29,6 @@ import com.liferay.oauth.util.OAuthAccessorConstants;
 import com.liferay.oauth.util.OAuthConsumer;
 import com.liferay.oauth.util.OAuthMessage;
 import com.liferay.oauth.util.OAuthValidator;
-import com.liferay.oauth.util.V10aOAuthDebugCacheListener;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.SingleVMPoolUtil;
 import com.liferay.portal.kernel.cluster.ClusterExecutorUtil;
